mv pipeline functions to seperate file
